{"schema":"decodex/commit/1","summary":"Split Radar request contracts","authority":"XY-1115"}
